+++ CHANGELOG.md
@@ -3,5 +3,12 @@ All notable changes to the "vsfirst" extension will be documented in this file.
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## 0.1.0 - 2017-01-14
+- added linux support
+- add diagnostic support for failed build
+- rewrite the Config class
+- automatically use the default installlation of Arduino IDE
+- only need to set the serial port setting if Arduino IDE installed with default folder on macOS and Windows
+
 ## [Unreleased]
 - Initial release